+++ src/lib/wapi/functions/revoke-invite-link.js
@@ -55,6 +55,6 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 export async function revokeGroupInviteLink(chatId) {
   var chat = Store.Chat.get(chatId);
   if (!chat.isGroup) return false;
-  await Store.GroupInvite.revokeGroupInvite(chat);
+  await Store.GroupInvite.sendRevokeGroupInviteCode(chat.id);
   return true;
 }