+++ src/lib/wapi/functions/set-group-description.js
@@ -52,41 +52,37 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNMMNMNMMMNMMNNMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNNNNMMNNNMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 */
-export enum InterfaceMode {
-  /**
-   * QR code page.
-   */
-  QR = 'QR',
-  /**
-   * Chat page.
-   */
-  MAIN = 'CONNECTION',
-  /**
-   * Loading page, waiting data from smartphone.
-   */
-  SYNCING = 'SYNCING',
-  /**
-   * Offline page, when there are no internet.
-   */
-  OFFLINE = 'OFFLINE',
-  /**
-   * Conflic page, when there are another whatsapp web openned.
-   */
-  CONFLICT = 'CONFLICT',
-  /**
-   * Blocked page, by proxy.
-   */
-  PROXYBLOCK = 'PROXYBLOCK',
-  /**
-   * Blocked page.
-   */
-  TOS_BLOCK = 'TOS_BLOCK',
-  /**
-   * Blocked page.
-   */
-  SMB_TOS_BLOCK = 'SMB_TOS_BLOCK',
-  /**
-   * Deprecated page.
-   */
-  DEPRECATED_VERSION = 'DEPRECATED_VERSION'
+
+/**
+ * Parameters to change group description
+ * @param {string} groupId group number
+ * @param {string} description group description
+ */
+export async function setGroupDescription(groupId, description) {
+  if (typeof description != 'string' || description.length === 0) {
+    return WAPI.scope(
+      undefined,
+      true,
+      null,
+      'It is necessary to write a text!'
+    );
+  }
+  const chat = await WAPI.sendExist(groupId);
+  if (chat && chat.status != 404) {
+    const m = { type: 'setGroupDescription', description };
+    const To = await WAPI.getchatId(chat.id);
+    return window.Store.GroupDesc.setGroupDesc(chat, description)
+      .then(() => {
+        const obj = WAPI.scope(To, false, 'OK', description);
+        Object.assign(obj, m);
+        return obj;
+      })
+      .catch(() => {
+        const obj = WAPI.scope(To, true, 'error', description);
+        Object.assign(obj, m);
+        return obj;
+      });
+  } else {
+    return chat;
+  }
 }